move some tests task collection
